refactor: extract shared RelationshipEdge.css.ts for overlays
